Increase Upgraders Used In RCL3 Rooms To 3 (#404)
* Increase Upgraders Used In RCL3 Rooms To 3

At Room Level 3 there cannot be a storage unit in the room (assuming it has not been downgraded).  Currently only 1 upgrader is utilised which makes progress form Room Level 3 to Room Level 4 slow.  As part of the efforts to improve the low level game behavior we can increase the target number of upgraders to help get the room up to RCL4 with storage as fast as possible.  

Currently the code defaults to 1 upgrader.  Room Controller Level 3 allows 10 extensions of 50 capacity and 1 spawn of 300 capacity, making the maximum spend on a single creep 800.  The current upgraders fixed parts cost 300 and each multi set costs 350, meaning the largest upgrader an RCL3 room can make will have 1 multi giving 5 work parts. That is 5 energy into an upgrader per tick.  

If we try to fully utilise one source of 3000 energy over 300 ticks, we need to transfer 10 energy per tick into the upgrader.  That will take 2 upgraders.  Most people have two sources in the room so we could supply 4 upgraders if all the energy went into them.  Let us make sure to leave something spare for building and so forth by making the target 3 upgraders.  

This should make it take only 33% of the time to reach room control level 4 in an ideal world.

* better upgrader maxCount calculation