Add Clear Spawn Queue Examples To commands.js (#613)
* Add Clear Spawn Queue Examples To commands.js

When the code gets stuck it is helpful to clear the command queues.  These examples will clear each level of queue in a room. 

This patch was mentioned in issue #525.

* Clear The Queue Stored In Memory

You must clear the queues stored in memory for it to have any real effect.

* Correct Example To spawnQueueMedium

If the queue was called spawnQueueMed at some point it is not now.  Use the present name of spawnQueueMedium.