Robust RCL5 & Smarter Hauler Spawn (#809)
* Spawn More Workers When There Is Work To Do

Level 5 rooms frequently collapse. They are at a sensitive point where the code base switches from non-storage to storage base behaviour and where links are introduced.  If the player does not sequence the construction of these facilities or if one is destroyed it can cause a room to struggle or fail.  

Rooms with long haulage distances from the energy source to the spawn/extension/storage also suffer greatly as we rely heavily on haulers and workers to do the transport of energy.  If there are no working link facilities the workers and haulers can spend all their time hauling energy and recharging the spawn.  In the worst cases they do this at a rate too low to support the continued generation or workers, haulers and miners.  

This change maintains the current behaviour of always having at least a value of one for the maximum number of workers.  This still leaves us open to that worker expiring and us relying on a hauler to energise a spawn.  However if there is no hauler in the room we will get a maximum number of workers of at least 2, which provides some protection for the times when one worker fades out.  Further we get another bonus worker for any existing construction sites (such as when there is a link to finish), which seems to help complete construction in my long haul rooms.  Finally we get another bonus worker if the room is fortifiable, this is nothing to do with the room fragility but it does stick with the idea of spending energy when there is excess.  

Using this code (and the other change I am about to add to hauler counts) my failing level 5 room recovered, completed its buildings (after 24 hours of making no progress on them) and is now thriving.  I am back down to 2 workers as I have excess energy to fortify the place.

* Spawn More Haulers When Drop Mining Energy

In long haul rooms energy is drop mined once the container by a distant spawn is full.  This causes workers and haulers to focus on moving energy.  The problem with this is that workers then do very little or no building work, as they spend all their time recharging distance spawns and extensions.  

To prevent this focus on charging in workers, we can spawn additional haulers whenever there is a significant (1000 in this case) amount of energy in the proximity of a spawn.  This should be enough help to give workers a chance to do some building, hopefully building of a pair of links to simplify the energy transfer and eliminate the need for the additional worker.  

With this change and the additional worker change I just introduced my struggling level 5 room was able to finally complete its building tasks (after 24 hours of having them stalled) and has begun to flourish.  Now the links are build there is no energy on the ground and I am back down the normal number of haulers.

* thinking of the wrong branch.